Move exp/runtime/catalog to api/runtime/catalog and reduce deps
