+++ .golangci.yml
@@ -146,12 +146,12 @@ linters:
           alias: runtimev1alpha1
         - pkg: sigs.k8s.io/cluster-api/api/runtime/v1beta2
           alias: runtimev1
+        - pkg: sigs.k8s.io/cluster-api/api/runtime/catalog
+          alias: runtimecatalog
         - pkg: sigs.k8s.io/cluster-api/api/runtime/hooks/v1alpha1
           alias: runtimehooksv1
         - pkg: sigs.k8s.io/cluster-api/exp/runtime/controllers
           alias: runtimecontrollers
-        - pkg: sigs.k8s.io/cluster-api/exp/runtime/catalog
-          alias: runtimecatalog
         - pkg: sigs.k8s.io/cluster-api/internal/runtime/client
           alias: internalruntimeclient
         - pkg: sigs.k8s.io/cluster-api/exp/runtime/client
