+++ internal/runtime/test/v1alpha2/groupversion_info.go
@@ -19,7 +19,7 @@ package v1alpha2
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	runtimecatalog "sigs.k8s.io/cluster-api/exp/runtime/catalog"
+	runtimecatalog "sigs.k8s.io/cluster-api/api/runtime/catalog"
 )
 
 var (
